+++ spec/scss_lint/linter/single_line_per_selector_spec.rb
@@ -109,8 +109,12 @@ describe SCSSLint::Linter::SingleLinePerSelector do
     it { should_not report_lint }
   end
 
-  context 'when selector is just an interpolated string' do
-    let(:css) { '#{$selector} {}' }
+  context 'when selector contains an interpolated string' do
+    let(:css) { <<-CSS }
+      div,
+      \#{$selector},
+      p {}
+    CSS
 
     it { should_not report_lint }
   end